labs/lab-02-doc-pipeline/requirements.txt
anthropic>=0.40.0
pydantic>=2.0
python-dotenv>=1.0.0
tenacity>=8.0
